Move the scan status initialisation code back out of the Scanner base class
Commit e7ccabf39ac33939920df31da7566b62e6094cc7 embodied the belief that,
as all the subclasses did the same thing here, this code could be moved to
the base class. Unfortunately, this belief was wrong: processors create
bare Scanners in order to find the rules they should apply! The overall
effect was to accidentally butcher the internal state of a scan every time
a processor crashed...