add explicit `on_delete` to all ForeignKeys and OneToOneFields
Django made this a hard requirement in 2.0 and later, for precisely
the reason seen here: all fields default to CASCADE, but that's
probably not what you want in quite a lot of cases. As an example,
should delete a scanner really affect all its reports as well?

I've gone through the various fields, and tried to apply the following
logic:

- Does the field allow `NULL` values? → SET_NULL
- Would deleting this necessarily affect a scan or report? → PROTECT
— Is this a detail in a more ‘important’ object? → CASCADE