scans: adjust stringification to be more meaningful
Previously, it was something like:

SCAN: Whatever

Now, it's:

Whatever — 2018-04-12 10:44


The idea is that any scan is identified by not only its name but also
the time at which it ran. Furthermore, explicitly stating the fact
that a scan is a scan, will, in most cases, be quite redundant; it
should be readily apparent from the context.

This primarily affects the admin UI.