scanner: infer whether a scan is running
Sometimes, a scanner can end up in an inconsistent state where no
actual scan is running, yet the scanner is marked as `is_running`.
Scans already have both a state a relation to the scanner, so the
`is_running` adds no extra information. Adding an attribute that
infers this information from our relational database should work
better in practice.

(The only downside I can think of is that this might perform somewhat
poorly with large databases; if so, we should add an index to the
state.)