filedomain: replace `check_mountpoint` with a simpler `is_mounted`
Since the change reverses the Boolean value of the result, I renamed
the method and changed it into a property. `os.path.ismount()` should
be fine for our purposes. According to the the docs,[1] the only
downside is that it “is not able to reliably detect bind mounts on the
same filesystem.” That's fine since we always mount proper network
shares.

[1] https://docs.python.org/3.7/library/os.path.html#os.path.ismount