properly prevent editing most fields on scanners
The proper way of preventing modifications to a field is to set it as
`disabled` in Django, so that Django disregards any modifications to
it. Merely setting it in the HTML will prevent the browser from
including it in the submission, but given that Django still expects
it, this effectively overwrites said value.

I've assumed this applies all fields on scanners other than the
following:

 - name
 - organisation
 - recipients
 - schedule