runners: don't suppress call to object.__init__()
This changes the logic to rely on the Method Resolution Order, or MRO,
rather than explicitly invoking two superclass initialisation methods.

When the superclass provides a given method, subclasses should call it
when overriding, even when it does nothing, as is the case with
`object.__init__()`.

As an example, given::

    #!/usr/bin/env python3

    class Base:
        def __init__(self):
            print('> BASE')
            super().__init__()
            print('< BASE')


    class Mixin:
        def __init__(self):
            print('> MIXIN')
            super().__init__()
            print('< MIXIN')


    class Derived(Base, Mixin):
        def __init__(self):
            print('> DERIVED')
            super().__init__()
            print('< DERIVED')


    Derived()


You get::

    > DERIVED
    > BASE
    > MIXIN
    < MIXIN
    < BASE
    < DERIVED