scanner manager: explicitly listen for child death
We use `asyncio` for this despite the added complexity. First of all,
AMQP is inherently asynchronous and fits this model rather well, and
second of all this suits the concurrent nature of the tasks well. In
particular, `asyncio` makes listening for dying children rather easy.

Now that we get an explicit notification when a child dies, I've
relaxed the regular check: Instead of every ten seconds, we run it
every 5 minutes.

https://redmine.magenta-aps.dk/issues/28795