Added a simple implementation of the WebSource.handles method
This is based on Requests rather than Scrapy, since the Twisted reactor
concept was a bit hard to fit into here; as a consequence, this code has to
implement more of the scraping process than I might have hoped. But it
searches for links, avoids revisiting URLs, and has a form of throttling,
so... that's not a bad start, is it?