Added a standard interface for retrieving the size of a FileResource
Like the last modified date, this is a thing that the rest of the engine
needs: unlike the last modified date, this has never had a common
interface, so every FileResource subclass needed different treatment. But
not any more!

(I /think/ this commit adds implementations of this method to all of the
relevant FileResource classes; this should really be tested somewhere...)