Changed the path abstraction behind Handles from pathlib.Path to str
pathlib.Path was corrupting HTTP URLs (and has already given us problems
with Zip file contents in the past), so this seemed like the best
solution. This was a relatively major change that was... actually quite
easy to do; almost everything was using functions on Handle and so didn't
care about the underlying representation of paths. Hooray encapsulation!