Added clarifying remarks to the Resource.get_size method's documentation
I wrote a test to make sure that the length of files and streams is equal
to the result of get_size(), but of course that's not actually true in the
HTTP world, where the header might give you the size of a *compressed*
object. So I'm updating the documentation instead!