Added the outline of the engine2-based pipeline to the repository
The time of great excitement is upon us! At the moment, the only thing
that's implemented is connecting to the AMQP broker: all of the components
will do that quite happily and will wait for orders that will never come
and for which no action is yet defined. But this is still an exciting
start, isn't it?