Ensured that normal use of SMBCSource doesn't leak smbc.Context references
Prior to this commit, a closed _SMBCFile would still contain a reference
to an smbc.File, which in turn contained a reference to the smbc.Context
produced by SMBCSource._generate_state. This reference was enough to stop
the GC from tearing that context down; in turn, this meant that SMB
connections to remote servers were never cleaned up, which in *turn* meant
that servers would quickly throttle or block us for being antisocial.
Grrrrrr!