Made all the existing SimpleRule subclasses handle None inputs gracefully
Here's a thought: the pipeline's processor stage goes to a lot of length to
hide processor values of None, which also has weird consequences for the
rest of the system (in particular, matches never get generated for an
object when this happens). But... why? Why can't we just protect rules
against this case and then let the matcher stage do its job?