Removed all of the SourceManager code that supported nesting
This code was useless at best and actively harmful at worst (which is why
commit d7d6f3b0b134551ae81533f8afd4241d04d06eae removed all attempts to use
it). Opening a Source essentially happens in reverse: you open a leaf,
which opens something further up, and so on:

  1. Open Zip file entry
    2. Open Zip file
      3. Open connection to remote server

This doesn't lend itself well to nesting -- creating a new SourceManager
for each layer of that makes it impossible to share the Zip file handle and
the network connection(!!) across all of the entries in the Zip file!